Move version check to crate
