+++ tools/pepsi/Cargo.toml
@@ -17,11 +17,9 @@ indicatif = { workspace = true }
 nao = { workspace = true }
 regex = { workspace = true }
 repository = { workspace = true }
-semver = { workspace = true }
-serde = { workspace = true }
 serde_json = { workspace = true }
 source_analyzer = { workspace = true }
 spl_network_messages = { workspace = true }
-toml = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
+versions = { workspace = true }
